Move ImageSlider to its own file
