+++ project/frontend/src/components/ImageSlider.css
@@ -1,0 +1,61 @@
+.slider {
+  position: relative;
+  width: 344px;
+  height: 200px;
+  overflow: hidden;
+  background-color: brown;
+}
+
+.slider-wrapper {
+  position: absolute;
+  white-space: nowrap;
+}
+
+.slider-wrapper.shifting {
+  transition: left 0.2s cubic-bezier(0.35, 0.91, 0.33, 0.97);
+}
+
+.slide {
+  display: inline-block;
+  width: 344px;
+  height: 200px;
+  cursor: grab;
+  text-align: center;
+  font-weight: bold;
+  font-size: 18px;
+  transition: all 1s;
+  position: relative;
+  background: #ffcf47;
+}
+
+.slide:active {
+  cursor: grabbing;
+}
+
+.control {
+  position: absolute;
+  top: 50%;
+  width: 24px;
+  height: 24px;
+  margin-top: -12px;
+  cursor: pointer;
+  z-index: 2;
+  transition: opacity 0.25s cubic-bezier(0.35, 0.91, 0.33, 0.97);
+}
+
+.control.hidden {
+  display: none;
+}
+
+.prev {
+  left: 8px;
+}
+
+.next {
+  right: 8px;
+}
+
+.prev:active,
+.next:active {
+  transform: scale(0.9);
+}
